+++ test_suite/tests/ui/string/validate/duplicated.rs
@@ -1,6 +1,6 @@
 use nutype::nutype;
 
-#[nutype(validate(min_len = 5, max_len = 255, min_len = 6))]
+#[nutype(validate(char_len_min = 5, char_len_max = 255, char_len_min = 6))]
 pub struct Email(String);
 
 fn main () {}